Добавлена проверка наличия проекта в БД при обновлении или удалении Добавлены unit тесты на команды и запросы для сущности проекта

+++ LogsMonitor.Application/Commands/Project/RemoveProjectCommand.cs
@@ -20,6 +20,13 @@ namespace LogsMonitor.Application.Commands
 
         public async Task Handle(RemoveProjectCommand request, CancellationToken cancellationToken)
         {
+            bool projectExists = await _projectRepository.Exists(request.ProjectId);
+
+            if (projectExists == false)
+            {
+                throw new NullReferenceException($"Проект с Id: {request.ProjectId} не найден");
+            }
+
             await _projectRepository.Remove(request.ProjectId);
         }
     }
